backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
langchain==0.2.2
chromadb==0.4.24
sentence-transformers==2.6.1
pdfminer.six==20231228
unstructured[local-inference]==0.12.6
tiktoken==0.6.0
pydantic==2.7.1
piper-tts==1.2.0      # free TTS